==> 01_data_transformations/10e_R_Seurat_PB_HSCMPP_integration.ipynb <==
# Seurat integration of PB HSC/MPP cells

library(Seurat)
library(future)
library(anndata)
library(batchelor)
library(scater)

plan(strategy = "multicore", workers = 4)

options(future.globals.maxSize = +Inf)

data.path = '2021/BloodPaper/h5ad/'

infile <- paste0(data.path, '20211203_COMBO10_PB_HSCMPP_counts.h5ad')

Sys.time()
data = read_h5ad(infile)
Sys.time()

data

Sys.time()

obj <- SingleCellExperiment( assays = List(counts = as(t(data$X), "CsparseMatrix")),
                              colData=data$obs,
                              rowData=data$var
                           )

Sys.time()

Sys.time()
obj = as.Seurat(obj, data = 'counts') 
Sys.time()

obj

---

---

data.rds = '2021/BloodPaper/rds/'

Sys.time()
saveRDS(file = paste0(data.rds, '20211203_COMBO_only_PB_HSCMPP_filtered_counts_Seurat3_input_obj.rds'), obj)
Sys.time()

rm(data)
rm(obj)

data.path = '2021/BloodPaper/'

data <- readRDS(file = paste0(data.path, 'rds/20211203_COMBO_only_PB_HSCMPP_filtered_counts_Seurat3_input_obj.rds') ) 


data.list <- SplitObject(object = data, split.by = "donor")


for (i in 1:length(x = data.list)) {
    data.list[[i]] <- NormalizeData(object = data.list[[i]], verbose = FALSE)
    data.list[[i]] <- FindVariableFeatures(object = data.list[[i]], 
        selection.method = "vst", nfeatures = 3000, verbose = FALSE)
}

reference.list <- data.list[ c("DOD1", "DOD2", "TQ198", "BP62j",
                               "BP37d", "BP74", "BP1c", "BP59h") ]

# find anchors                               
Sys.time()
data.anchors <- FindIntegrationAnchors(object.list = reference.list, dims = 1:15
                                        )
Sys.time()

saveRDS(file = paste0(data.path, 'rds/20211203_COMBO_only_PB_HSCMPP_Seurat3_VST_classic_anchors.rds'), data.anchors)        

# integrate features
Sys.time()
data.integrated <- IntegrateData(anchorset = data.anchors, dims = 1:15)
Sys.time()

saveRDS(file = paste0(data.path, 'rds/20211203_COMBO_only_PB_HSCMPP_Seurat3_VST_classic_integrated.rds'), data.integrated)

#Making sure cell/gene orders are the same
data.integrated = data.integrated[row.names(data), colnames(data)]

#Convertin to AnnData
x = data.integrated@assays$integrated@data
x = AnnData(X = t(x),
            obs = data.integrated@meta.data
           )

# Export/save anndata

outfile <-  paste0(data.path, 'h5ad/20211203_COMBO_only_PB_HSCMPP_Seurat3_VST_classic_integrated.h5ad')     

Sys.time()
x$write_h5ad(outfile, compression = 'lzf')
Sys.time()
